feat: separate 'install'
